Minor improvement to delete DB script

+++ Demonstrator/ResearchInterface/cleandb.sh
@@ -1,3 +1,8 @@
 # Reinitialize db
-rm db.sqlite3 
+if [ -f db.sqlite3 ]
+then
+    echo "Delete database"
+    rm db.sqlite3
+fi
+
 find . -name "*000*.py" -exec rm {} \;
